+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge_board/ChallengeBoardAdapter.kt
@@ -1,0 +1,4 @@
+package com.side.runwithme.view.challenge_board
+
+class ChallengeBoardAdapter {
+}
