+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge_create/ChallengeCreateActivity.kt
@@ -1,4 +1,4 @@
-package com.side.runwithme.view.challenge_list.create
+package com.side.runwithme.view.challenge_create
 
 import com.example.seobaseview.base.BaseActivity
 import com.side.runwithme.R
