+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge_create/ChallengeCreateBindingAdapter.kt
@@ -1,4 +1,4 @@
-package com.side.runwithme.view.challenge_list.create
+package com.side.runwithme.view.challenge_create
 
 import android.view.View
 import android.widget.TextView
