+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge_create/ChallengeCreateStep2Fragment.kt
@@ -1,4 +1,4 @@
-package com.side.runwithme.view.challenge_list.create
+package com.side.runwithme.view.challenge_create
 
 import android.app.DatePickerDialog
 import androidx.fragment.app.activityViewModels
@@ -6,7 +6,7 @@ import androidx.navigation.fragment.findNavController
 import com.example.seobaseview.base.BaseFragment
 import com.side.runwithme.R
 import com.side.runwithme.databinding.FragmentChallengeCreate2Binding
-import com.side.runwithme.view.challenge_list.create.dialog.GoalWeeksDialog
+import com.side.runwithme.view.challenge_create.dialog.GoalWeeksDialog
 import dagger.hilt.android.AndroidEntryPoint
 import java.util.Calendar
