+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge_create/ChallengeCreateViewModel.kt
@@ -1,4 +1,4 @@
-package com.side.runwithme.view.challenge_list.create
+package com.side.runwithme.view.challenge_create
 
 import android.net.Uri
 import android.util.Log
