+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge_create/dialog/CostDialog.kt
@@ -1,4 +1,4 @@
-package com.side.runwithme.view.challenge_list.create.dialog
+package com.side.runwithme.view.challenge_create.dialog
 
 import android.os.Build
 import android.widget.NumberPicker
@@ -8,7 +8,7 @@ import com.example.seobaseview.base.BaseDialogFragment
 import com.side.runwithme.R
 import com.side.runwithme.databinding.DialogCostBinding
 import com.side.runwithme.util.repeatOnStarted
-import com.side.runwithme.view.challenge_list.create.ChallengeCreateViewModel
+import com.side.runwithme.view.challenge_create.ChallengeCreateViewModel
 
 class CostDialog(): BaseDialogFragment<DialogCostBinding>(
     R.layout.dialog_cost) {
