+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge_create/dialog/MaxMemberDialog.kt
@@ -1,4 +1,4 @@
-package com.side.runwithme.view.challenge_list.create.dialog
+package com.side.runwithme.view.challenge_create.dialog
 
 import android.os.Build
 import android.widget.NumberPicker
@@ -6,11 +6,12 @@ import androidx.core.content.ContextCompat
 import androidx.fragment.app.activityViewModels
 import com.example.seobaseview.base.BaseDialogFragment
 import com.side.runwithme.R
-import com.side.runwithme.databinding.DialogGoalWeeksBinding
+import com.side.runwithme.databinding.DialogMaxMemberBinding
 import com.side.runwithme.util.repeatOnStarted
-import com.side.runwithme.view.challenge_list.create.ChallengeCreateViewModel
+import com.side.runwithme.view.challenge_create.ChallengeCreateViewModel
 
-class GoalWeeksDialog: BaseDialogFragment<DialogGoalWeeksBinding>(R.layout.dialog_goal_weeks) {
+class MaxMemberDialog(): BaseDialogFragment<DialogMaxMemberBinding>(
+    R.layout.dialog_max_member) {
 
     private val challengeCreateViewModel by activityViewModels<ChallengeCreateViewModel>()
 
@@ -23,12 +24,11 @@ class GoalWeeksDialog: BaseDialogFragment<DialogGoalWeeksBinding>(R.layout.dialo
         initNumberPicker()
 
         initViewModelCallbacks()
-
     }
 
     private fun initViewModelCallbacks(){
         repeatOnStarted {
-            challengeCreateViewModel.dialogGoalWeeksEventFlow.collect {
+            challengeCreateViewModel.dialogMaxMemberEventFlow.collect {
                 if(it is ChallengeCreateViewModel.Event.Success){
                     dismiss()
                 }
@@ -38,15 +38,14 @@ class GoalWeeksDialog: BaseDialogFragment<DialogGoalWeeksBinding>(R.layout.dialo
 
     private fun initNumberPicker(){
         binding.apply {
-            np.minValue = 1
-            np.maxValue = 25
-            np.value = 4
+            np.minValue = 2
+            np.maxValue = 20
+            np.value = 7
+            np.wrapSelectorWheel = false
+
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
                 np.textColor = ContextCompat.getColor(requireActivity(), R.color.main_blue)
             }
-
-            //순환 안되게 막기
-            np.wrapSelectorWheel = false
 
             np.descendantFocusability = NumberPicker.FOCUS_BLOCK_DESCENDANTS
         }
