+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge_detail/ChallengeDetailBindingAdapter.kt
@@ -1,6 +1,5 @@
-package com.side.runwithme.view.challenge_list.detail
+package com.side.runwithme.view.challenge_detail
 
-import android.content.res.ColorStateList
 import android.view.View
 import android.widget.TextView
 import androidx.appcompat.widget.AppCompatButton
