+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge_detail/ChallengeDetailFragment.kt
@@ -1,4 +1,4 @@
-package com.side.runwithme.view.challenge_list.detail
+package com.side.runwithme.view.challenge_detail
 
 import android.content.Intent
 import androidx.fragment.app.viewModels
