--- RunWithMe/presentation/src/main/java/com/side/runwithme/view/challenge_list/detail/board/ChallengeBoardAdapter.kt
@@ -1,4 +1,0 @@
-package com.side.runwithme.view.challenge_list.detail.board
-
-class ChallengeBoardAdapter {
-}
