+++ RunWithMe/presentation/src/main/res/layout/fragment_challenge_detail.xml
@@ -4,14 +4,14 @@
     <data>
         <variable
             name="challengeDetailVM"
-            type="com.side.runwithme.view.challenge_list.detail.ChallengeDetailViewModel" />
+            type="com.side.runwithme.view.challenge_detail.ChallengeDetailViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.challenge_list.detail.ChallengeDetailFragment">
+        tools:context=".view.challenge_detail.ChallengeDetailFragment">
 
         <com.side.runwithme.component.CustomToolbar
             android:id="@+id/toolbar_main"
